add custom hook for media fetching and update Home component to use it
