+++ src/views/Home.jsx
@@ -1,50 +1,11 @@
-import {useEffect, useState} from 'react';
 import MediaRow from '../components/MediaRow';
 import SingleView from '../components/SingleView';
-import {fetchData} from '../utils/fetchData';
-import {uniqBy} from 'lodash';
+import useMedia from '../hooks/apiHooks.js';
+import {useState} from 'react';
 
 const Home = () => {
-  const [mediaArray, setMediaArray] = useState([]);
+  const {mediaArray} = useMedia();
   const [selectedItem, setSelectedItem] = useState(null);
-  console.log('selectedItem', selectedItem);
-
-  const getMedia = async () => {
-    try {
-      const mediaData = await fetchData(
-        import.meta.env.VITE_MEDIA_API + '/media',
-      );
-
-      const uniqueUserIds = uniqBy(mediaData, 'user_id');
-
-      console.log('uniqueUserIds', uniqueUserIds);
-
-      const authApiUrl = import.meta.env.VITE_AUTH_API;
-
-      const userData = await Promise.all(
-        uniqueUserIds.map(
-          async (item) =>
-            await fetchData(`${authApiUrl}/users/${item.user_id}`),
-        ),
-      );
-
-      console.log('userData', userData);
-
-      const newData = mediaData.map((item) => {
-        const user = userData.find(({user_id}) => user_id === item.user_id);
-        return {...item, username: user.username};
-      });
-
-      setMediaArray(newData);
-    } catch (error) {
-      console.error('error', error);
-    }
-  };
-
-  useEffect(() => {
-    getMedia();
-  }, []);
-
   console.log('mediaArray', mediaArray);
 
   return (
